+++ GraphOverflow/GraphOverflow.Dto/IPostDto.cs
@@ -7,6 +7,6 @@ namespace GraphOverflow.Dtos
     public int Id { get; set; }
     public string Content { get; set; }
     public DateTime CreatedAt { get; set; }
-    public int UpVotes { get; set; }
+    public long UpVotes { get; set; }
   }
 }
